feat: automap mini (#62)
* feat: fog-of-war foundation phase 1 complete

* feat: top-right panel consolidation and mini map widget phase 2 complete

* feat: full-screen automap overlay phase 3 complete

* feat: poi markers and legend phase 4 complete

* feat: config, save/load verification, and sdk integration phase 5 complete

* feat: automap mostly done

* feat: automap clean up

+++ sdk/campaign_builder/src/map_editor.rs
@@ -2305,7 +2305,7 @@ impl EventEditorState {
                     None
                 };
                 let key_item_id = if self.furniture_type == FurnitureType::Door {
-                    self.furniture_key_item_id.trim().parse::<u32>().ok()
+                    self.furniture_key_item_id.trim().parse::<u8>().ok()
                 } else {
                     None
                 };
